Revert "Fixes #57"
This reverts commit d26e12a1ff6f14ba7a000043a0a02d8e788289c0.

+++ encog-core-cs/App/Analyst/Analyze/AnalyzedField.cs
@@ -140,7 +140,7 @@ namespace Encog.App.Analyst.Analyze
             {
                 try
                 {
-                    Int64 i = Int64.Parse(str);
+                    int i = Int32.Parse(str);
                     Max = Math.Max(i, Max);
                     Min = Math.Min(i, Min);
                     if (!accountedFor)
